requirements.txt
pytesseract
pillow
python-telegram-bot
mss
rapidfuzz